Oops.  That's not what I meant.
(Though, yes, we'd love for a new developer to "complete" us.  To join
our ethereal thread of destiny; to become so inseparable from the
mythological Team that it is impossible to see the boundaries, aside
from the occasional fleeting glimpse of the cuff of one's jeans.  To
sip eagerly from the tea in which steeped the crushed souls of those
who Developed Before.  Crushed in, you know, a good way.)